Split federationparams.cpp into CDynamicParams/genesis-block lifecycle
FederationParams used to own both static per-deployment identity
(networkId, message-start bytes, softfork registration) and the
per-run genesis-block-reading/CXFieldHistory-seeding lifecycle, which
forced a circular dependency: FederationParams needed key_io.cpp
(DecodeDestination) for genesis parsing, and key_io.cpp needed
FederationParams for network identity.

Split the genesis lifecycle out into a new CDynamicParams class
(src/dynamicparams.{cpp,h}) and genesis-block creation into
src/genesisblock.{cpp,h}, both reached via SelectDynamicParams()/
DynamicParams() rather than FederationParams().GenesisBlock()/
CreateFederationParams(mode, loadGenesis). Updates every caller
(chainstate.cpp, file_io.cpp, init.cpp, interfaces/node.cpp,
rpc/rawtransaction.cpp, the tapyrus-cli/genesis/verify entrypoints,
and the unit tests) to the new API, plus two missing-include fixes
(xfieldhistory.h, bench/bench.cpp) needed for the moved types to
actually compile.
