+++ src/bench/bench.cpp
@@ -3,6 +3,7 @@
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
 #include <bench/bench.h>
+#include <dynamicparams.h>
 #include <federationparams.h>
 #include <assert.h>
 #include <iostream>
