+++ src/genesisblock.h
@@ -1,0 +1,21 @@
+// Copyright (c) 2010 Satoshi Nakamoto
+// Copyright (c) 2009-2018 The Bitcoin Core developers
+// Copyright (c) 2019-2021 Chaintope Inc.
+// Distributed under the MIT software license, see the accompanying
+// file COPYING or http://www.opensource.org/licenses/mit-license.php.
+
+#ifndef BITCOIN_GENESISBLOCK_H
+#define BITCOIN_GENESISBLOCK_H
+
+#include <primitives/block.h>
+#include <pubkey.h>
+#include <key.h>
+
+#include <string>
+
+/**
+ * @returns a signed genesis block.
+ */
+CBlock createGenesisBlock(const CPubKey& aggregatePubkey, const CKey& privateKey, const time_t blockTime=time(0), const std::string paytoAddress="");
+
+#endif // BITCOIN_GENESISBLOCK_H
