+++ src/init.cpp
@@ -13,6 +13,7 @@
 #include <chain.h>
 #include <chainparams.h>
 #include <checkpoints.h>
+#include <dynamicparams.h>
 #include <consensus/validation.h>
 #include <fs.h>
 #include <httpserver.h>
@@ -1255,18 +1256,19 @@ bool AppInitMain()
     //Read genesis block from file now - we are sure that data dir exists.
     try {
         SelectFederationParams(gArgs.GetChainMode());
+        SelectDynamicParams(ReadGenesisBlock());
     } catch (const std::exception& e) {
         fprintf(stderr, "Error: %s\n", e.what());
         return false;
     }
-    LogPrintf("Genesis Block [%s] of Tapyrus network [%s] Loaded successfully\n", FederationParams().GenesisBlock().GetHash().ToString(), FederationParams().NetworkIDString());
+    LogPrintf("Genesis Block [%s] of Tapyrus network [%s] Loaded successfully\n", DynamicParams().GenesisBlock().GetHash().ToString(), FederationParams().NetworkIDString());
 
     // Initialise ChainTxData from the genesis block timestamp so that
     // GuessVerificationProgress produces meaningful values during IBD.
     // nTime is wall-clock time (so nNow - nTime ≈ 0 and fTxTotal ≈ nTxCount),
     // and nTxCount is the expected total transactions from genesis to now.
-    if (!FederationParams().GenesisBlock().vtx.empty()) {
-        const int64_t genesisTime = FederationParams().GenesisBlock().nTime;
+    if (!DynamicParams().GenesisBlock().vtx.empty()) {
+        const int64_t genesisTime = DynamicParams().GenesisBlock().nTime;
         const double dTxRate = 1.0 / Params().GetConsensus().nExpectedBlockTime;
         const int64_t nNow = time(nullptr);
         const int64_t nTxCount = std::max(int64_t(1),
@@ -1499,7 +1501,7 @@ bool AppInitMain()
 
                 // If the loaded chain has a wrong genesis, bail out immediately
                 // (we're likely using a testnet datadir, or the other way around).
-                if (!mapBlockIndex.empty() && !LookupBlockIndex(FederationParams().GenesisBlock().GetHash())) {
+                if (!mapBlockIndex.empty() && !LookupBlockIndex(DynamicParams().GenesisBlock().GetHash())) {
                     return InitError(_("Incorrect or no genesis block found. Wrong datadir for network?"));
                 }
 
@@ -1520,7 +1522,7 @@ bool AppInitMain()
                 } else {
                     // Initialize xfieldHistory here as it is not initialized
                     // when genesis block is loaded from the file during reindex
-                    CXFieldHistory history(FederationParams().GenesisBlock());
+                    CXFieldHistory history(DynamicParams().GenesisBlock());
                 }
 
                 // At this point we're either in reindex or we've loaded a useful
